requirements.txt
keras
numpy
pandas
python-chess
torch
